Prevent measuring request duration metric for invalid JSON-RPC endpoints
